refactor: restructured dirs
