+++ packages/react-app/src/routes/create/Create.jsx
@@ -37,16 +37,16 @@ import React, { useEffect, useState } from "react";
 import { toWei } from "web3-utils";
 import { useFieldArray, useForm } from "react-hook-form";
 import { useHistory } from "react-router-dom";
-import CenteredFrame from "../components/layout/CenteredFrame";
-import dips from "../dips";
-import AddressInputChakra from "../components/AddressInputChakra";
-import AddressChakra from "../components/AddressChakra";
-import ElectionCard from "../components/Cards/ElectionCard";
-import { blockExplorer } from "../App";
+import CenteredFrame from "../../components/layout/CenteredFrame";
+import dips from "../../dips";
+import AddressInputChakra from "../../components/AddressInputChakra";
+import AddressChakra from "../../components/AddressChakra";
+import ElectionCard from "../../components/Cards/ElectionCard";
+import { blockExplorer } from "../../App";
 import { ethers } from "ethers";
 import { Form, Card, InputNumber, Space } from "antd";
-import MongoDBController from "../controllers/mongodbController";
-import { CERAMIC_PREFIX } from "../dips/helpers";
+import MongoDBController from "../../controllers/mongodbController";
+import { CERAMIC_PREFIX } from "../../dips/helpers";
 
 const CURRENCY = process.env.REACT_APP_NETWORK_SYMBOL;
 const TOKEN = process.env.REACT_APP_TOKEN_SYMBOL;
