+++ packages/react-app/src/routes/index.js
@@ -1,0 +1,3 @@
+export { default as Home } from "./home/Home";
+export { default as Create } from "./create/Create";
+export { default as Party } from "./party/Party";
