--- packages/react-app/src/views/index.js
@@ -1,3 +1,0 @@
-export { default as Home } from "./Home";
-export { default as Create } from "./Create";
-export { default as Election } from "./Election";
